hopefully stop any runtimes

+++ code/mob/living/critter.dm
@@ -347,7 +347,9 @@ ABSTRACT_TYPE(/mob/living/critter)
 		src.ghostize()
 		qdel (src)
 
-	proc/remove_arm(var/mob/living/critter/M, var/arm)
+	proc/remove_arm(var/mob/living/critter/M, var/arm) // for removing the arms of brullbars and bears
+		if(!arm)
+			return
 		if(arm == "left")
 			var/datum/handHolder/HH = hands[1]
 			qdel(HH)
